Providers and Products Completed

+++ app/Http/Controllers/Insurance/InsuranceProductController.php
@@ -7,7 +7,6 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\Insurance\ProviderAndProducts\InsuranceProductRequest;
 use App\Models\Core\CodeDetail;
 use App\Services\Insurance\ProviderAndProducts\InsuranceProductService;
-use Illuminate\Http\Request;
 use App\Models\Insurance\InsuranceProduct;
 use App\Models\Insurance\InsuranceProvider;
 use App\Enums\Core\PermissionEnum;
@@ -108,6 +107,11 @@ class InsuranceProductController extends Controller
         $this->authorize(PermissionEnum::InsuranceProductDelete, InsuranceProduct::class);
         try {
             $product = InsuranceProduct::findOrFail($Id);
+
+            if ($product->policies()->exists()) {
+                return redirect()->back()
+                ->withErrors(['error' => 'This Product is in use and cannot be deleted.']);
+            } 
             $product->delete();
 
             return redirect()->route('bancassurance.products.index')
@@ -120,40 +124,5 @@ class InsuranceProductController extends Controller
                 ->withInput();
         }
     }
-
-    // public function mapForm($id)
-    // {
-    //     $product = DB::table('t_InsuranceProducts')->find($id);
-
-    //     $providers = DB::table('t_InsuranceProviders')->pluck('Name', 'Id')->toArray();
-
-    //     $policyTypes = DB::table('t_CodeDetails')
-    //         ->where('CodeID', 'POLICY_TYPE')
-    //         ->pluck('Description', 'Id')
-    //         ->toArray();
-
-    //     return view('bancassurance.products.map', compact('product', 'providers', 'policyTypes'));
-    // }
-
-    // public function storeMap(Request $request, $id)
-    // {
-    //     $request->validate([
-    //         'InsuranceProviderID' => 'required|exists:t_InsuranceProviders,Id',
-    //         'PolicyTypeID' => 'nullable|exists:t_CodeDetails,Id',
-    //         'CustomName' => 'nullable|string|max:150',
-    //         'CommissionType' => 'required|in:Flat,Tiered',
-    //     ]);
-
-    //     DB::table('t_InsuranceProviderProducts')->insert([
-    //         'ProductID' => $id,
-    //         'InsuranceProviderID' => $request->InsuranceProviderID,
-    //         'PolicyTypeID' => $request->PolicyTypeID,
-    //         'CustomName' => $request->CustomName,
-    //         'CommissionType' => $request->CommissionType,
-    //         'CreatedAt' => now(),
-    //     ]);
-
-    //     return redirect()->route('bancassurance.products.index')->with('success', 'Product mapped to provider successfully.');
-    // }
 
 }
